Some edits to YAAM code
Changed OverrideArtCacheSize to set the art cache size to 261.
Fixed typo in ddraw.ini.

+++ sfall/Modules/DamageMod.h
@@ -30,7 +30,7 @@ public:
 
 	static int formula;
 	static void DamageGlovz(fo::ComputeAttackResult &ctd, DWORD &accumulatedDamage, long rounds, long armorDT, long armorDR, long bonusRangedDamage, long multiplyDamage, long difficulty);
-	static void DamageYAAM(fo::ComputeAttackResult &ctd, DWORD &accumulatedDamage, int rounds, int armorDT, int armorDR, int bonusRangedDamage,int multiplyDamage, int difficulty);
+	static void DamageYAAM(fo::ComputeAttackResult &ctd, DWORD &accumulatedDamage, long rounds, long armorDT, long armorDR, long bonusRangedDamage, long multiplyDamage, long difficulty);
 
 	static long GetHtHMinDamageBonus(fo::GameObject* source);
 };
